# backend/docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ethnic Fashion SaaS Backend API",
    "version": "1.0.0",
    "description": "Multi-tenant backend API with JWT auth, RBAC, and subscription feature controls."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "OkResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string", "example": "ok" },
          "data": {}
        },
        "required": ["success", "data"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "code": { "type": "string", "example": "VALIDATION_ERROR" },
          "message": { "type": "string", "example": "Request validation failed" }
        },
        "required": ["success", "code", "message"]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 8 }
        },
        "required": ["email", "password"]
      },
      "RefreshRequest": {
        "type": "object",
        "properties": {
          "refreshToken": { "type": "string", "minLength": 20 }
        },
        "required": ["refreshToken"]
      },
...
      "PlanCreate": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 2 },
          "code": { "type": "string", "minLength": 2, "maxLength": 50 },
          "description": { "type": "string", "maxLength": 2000 },
          "billingCycle": { "type": "string", "enum": ["MONTHLY", "QUARTERLY", "YEARLY"] },
          "price": { "type": "number", "minimum": 0 },
          "currency": { "type": "string", "minLength": 3, "maxLength": 3 },
          "isActive": { "type": "boolean" },
          "features": { "type": "array", "items": { "type": "string" } },
          "limits": { "type": "object", "additionalProperties": true }
        },
        "required": ["name", "code", "billingCycle", "price"]
      }
    }
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login with email/password",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } }
        },
        "responses": {
          "200": { "description": "Login successful", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OkResponse" } } } },
          "400": { "description": "Validation failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Invalid credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Rotate refresh token",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefreshRequest" } } }
        },
        "responses": {
          "200": { "description": "Token refreshed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OkResponse" } } } },
          "401": { "description": "Refresh token invalid/revoked", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Revoke refresh token session",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefreshRequest" } } }
        },
        "responses": {
          "200": { "description": "Logged out" }
        }
      }
    },
...
    "/api/v1/subscriptions/plans": {
      "get": {
        "tags": ["Subscriptions"],
        "summary": "List subscription plans",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "activeOnly", "in": "query", "schema": { "type": "boolean", "default": false } }],
        "responses": { "200": { "description": "Plan list" } }
      },
      "post": {
        "tags": ["Subscriptions"],
        "summary": "Create subscription plan (SUPER_ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PlanCreate" } } } },
        "responses": { "201": { "description": "Plan created" } }
      }
    },
    "/api/v1/subscriptions/plans/{planId}": {
      "patch": {
        "tags": ["Subscriptions"],
        "summary": "Update subscription plan",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "planId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PlanCreate" } } } },
        "responses": { "200": { "description": "Plan updated" } }
      },
      "delete": {
        "tags": ["Subscriptions"],
        "summary": "Deactivate subscription plan",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "planId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Plan deactivated" } }
      }
    },
    "/api/v1/subscriptions/organizations/{organizationId}/current": {
      "get": {
        "tags": ["Subscriptions"],
        "summary": "Get organization current subscription",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "organizationId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Current subscription" } }
      },
      "put": {
        "tags": ["Subscriptions"],
        "summary": "Assign plan to organization",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "organizationId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "201": { "description": "Assigned" } }
      },
      "patch": {
        "tags": ["Subscriptions"],
        "summary": "Update organization current subscription",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "organizationId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Updated" } }
      }
    },
    "/api/v1/subscriptions/organizations/{organizationId}/current/cancel": {
      "post": {
        "tags": ["Subscriptions"],
        "summary": "Cancel organization current subscription",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "organizationId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Canceled" } }
      }
    },
    "/api/v1/subscriptions/me/current": {
      "get": {
        "tags": ["Subscriptions"],
        "summary": "Get current organization subscription for current user",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Current subscription" } }
      }
    },
    "/api/v1/subscriptions/me/features/{featureKey}": {
      "get": {
        "tags": ["Subscriptions"],
        "summary": "Check current organization feature access",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "featureKey", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Feature access result" } }
      }
    }
  }
}